+++ src/mako/http/response/Cookies.php
@@ -120,7 +120,7 @@ class Cookies implements Countable, IteratorAggregate
 	}
 
 	/**
-	 * Adds a signed cookie.
+	 * Adds a raw unsigned cookie.
 	 *
 	 * @param  string                      $name    Cookie name
 	 * @param  string                      $value   Cookie value
@@ -128,14 +128,43 @@ class Cookies implements Countable, IteratorAggregate
 	 * @param  array                       $options Cookie options
 	 * @return \mako\http\response\Cookies
 	 */
-	public function addSigned(string $name, string $value, int $ttl = 0, array $options = []): Cookies
+	public function addRaw(string $name, string $value, int $ttl = 0, array $options = []): Cookies
+	{
+		return $this->add($name, $value, $ttl, $options, true);
+	}
+
+	/**
+	 * Adds a signed cookie.
+	 *
+	 * @param  string                      $name    Cookie name
+	 * @param  string                      $value   Cookie value
+	 * @param  int                         $ttl     Time to live - if omitted or set to 0 the cookie will expire when the browser closes
+	 * @param  array                       $options Cookie options
+	 * @param  bool                        $raw     Set the cookie without urlencoding the value?
+	 * @return \mako\http\response\Cookies
+	 */
+	public function addSigned(string $name, string $value, int $ttl = 0, array $options = [], bool $raw = false): Cookies
 	{
 		if(empty($this->signer))
 		{
 			throw new RuntimeException('A [ Signer ] instance is required to sign cookies.');
 		}
 
-		return $this->add($name, $this->signer->sign($value), $ttl, $options);
+		return $this->add($name, $this->signer->sign($value), $ttl, $options, $raw);
+	}
+
+	/**
+	 * Adds a raw signed cookie.
+	 *
+	 * @param  string                      $name    Cookie name
+	 * @param  string                      $value   Cookie value
+	 * @param  int                         $ttl     Time to live - if omitted or set to 0 the cookie will expire when the browser closes
+	 * @param  array                       $options Cookie options
+	 * @return \mako\http\response\Cookies
+	 */
+	public function addRawSigned(string $name, string $value, int $ttl, array $options = []): Cookies
+	{
+		return $this->addSigned($name, $value, $ttl, $options, true);
 	}
 
 	/**